john branch into dev

See merge request ibs-repo/csc309-winter-2023/group_634/p3!3

+++ easy_chef/backend/recipes/views.py
@@ -5,7 +5,8 @@ from rest_framework.permissions import *
 from rest_framework.generics import *
 from rest_framework.views import *
 from rest_framework import status
-from accounts.models import MyUser
+from accounts.models import MyUser, ShoppingList
+from accounts.views import IsTokenValid
 from recipes.serializers import *
 from recipes.models import *
 
@@ -29,12 +30,17 @@ from recipes.serializers import CuisineSerializer, DietSerializer, \
 # Create your views here.
 
 class InstructionFileUpdate(APIView):
+    permission_classes = [IsAuthenticated, IsTokenValid]
+
     def post(self, request, instruction_id):
         # recipe_id = self.kwargs['recipe_id']
         try:
             instruction = Instruction.objects.get(id=instruction_id)
         except Instruction.DoesNotExist:
             return Response({'instruction_id': 'instruction does not exist.'}, status=404)
+
+        if instruction.recipe.owner != request.user:
+            return Response({'instruction_id': 'instruction does not belong to user.'}, status=403)
 
         instruction.photos_or_videos.all().delete()
         files = request.FILES
@@ -51,12 +57,17 @@ class InstructionFileUpdate(APIView):
 
 
 class RecipeFileUpdate(APIView):
+    permission_classes = [IsAuthenticated, IsTokenValid]
+
     def post(self, request, recipe_id):
         # recipe_id = self.kwargs['recipe_id']
         try:
             recipe = Recipe.objects.get(id=recipe_id)
         except Recipe.DoesNotExist:
             return Response({'recipe_id': 'Recipe does not exist.'}, status=404)
+        if recipe.owner != request.user:
+            return Response({'recipe_id': 'Recipe does not belong to user.'}, status=403)
+
         recipe.photos_or_videos.all().delete()
         files = request.FILES
         if files:
@@ -75,7 +86,7 @@ class RecipeFileUpdate(APIView):
 
 
 class UpdateRecipe(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsTokenValid]
 
     def post(self, request, recipe_id):
         try:
@@ -83,6 +94,7 @@ class UpdateRecipe(APIView):
         except Recipe.DoesNotExist:
             return Response({'recipe_id': 'Recipe does not exist.'}, status=404)
         errors = {}
+
         if recipe.owner != request.user:
             return Response({'recipe_id': 'Recipe does not belong to user.'}, status=403)
 
@@ -232,7 +244,7 @@ class UpdateRecipe(APIView):
 
 
 class CreateView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsTokenValid]
 
     def post(self, request):
         errors = {}
@@ -376,17 +388,20 @@ class CreateView(APIView):
 
 
 class RecipeFileUpload(APIView):
+    permission_classes = [IsAuthenticated, IsTokenValid]
+
     def post(self, request, recipe_id):
         # recipe_id = self.kwargs['recipe_id']
         try:
             recipe = Recipe.objects.get(id=recipe_id)
         except Recipe.DoesNotExist:
             return Response({'recipe_id': 'Recipe does not exist.'}, status=404)
+        if recipe.owner != request.user:
+            return Response({'user': 'You are not the owner of this recipe.'}, status=403)
         # recipe.photos_or_videos.all().delete()
         files = request.FILES
-        print(request)
         if files:
-            print(request)
+
             file_dict = MultiValueDict(files)
             for file_key in file_dict.keys():
                 file_list = file_dict.getlist(file_key)
@@ -405,12 +420,16 @@ class RecipeFileUpload(APIView):
 
 
 class InstructionFileUpload(APIView):
+    permission_classes = [IsAuthenticated, IsTokenValid]
+
     def post(self, request, instruction_id):
         # recipe_id = self.kwargs['recipe_id']
         try:
             instruction = Instruction.objects.get(id=instruction_id)
         except Instruction.DoesNotExist:
             return Response({'instruction_id': 'instruction does not exist.'}, status=404)
+        if instruction.recipe.owner != request.user:
+            return Response({'user': 'You are not the owner of this recipe.'}, status=403)
         # file = request.FILES['file']
         # name = file.name
         # instruction_file = InstructionFile.objects.create(name=name, recipe=instruction, file=file)
@@ -451,12 +470,19 @@ class IngredientSearchView(View):
 
 
 class DeleteRecipe(APIView):
+    permission_classes = [IsAuthenticated, IsTokenValid]
+
     def delete(self, request):
+
         recipe_id = request.data.get('recipe_id', '')
         try:
             recipe = Recipe.objects.get(id=recipe_id)
         except Recipe.DoesNotExist:
             return Response({'error': 'Recipe not found.'}, status=404)
+
+        if recipe.owner != request.user:
+            return Response({'recipe_id': 'Recipe does not belong to user.'}, status=403)
+
         recipe.delete()
         return Response({'success message': 'Recipe was deleted from the database'}, status=204)
 
@@ -464,24 +490,106 @@ class DeleteRecipe(APIView):
 class RecipeDetails(APIView):
     """
     This view is used to get the details of a recipe.
-    GET /recipes/recipe-details/
-    Payloads:
-        - recipe_name
+    GET /recipes/recipe-details/<int:recipe_id>/
     Returns: Recipe details as specified in model
     """
 
-    def get(self, request):
-        recipe_name = request.GET.get('recipe_name', None)
-        if recipe_name is None:
-            return Response({'error': 'Recipe name is required'}, status=status.HTTP_400_BAD_REQUEST)
-        recipe = Recipe.objects.filter(name=recipe_name)
-        if len(recipe) == 0:
-            return Response({'error': 'Recipe not found'}, status=status.HTTP_404_NOT_FOUND)
-        serializer = RecipeSerializer(recipe[0])
-        return Response(serializer.data, status=status.HTTP_200_OK)
+    def get(self, request,recipe_id):
+        # recipe_name = request.GET.get('recipe_name', None)
+        # if recipe_name is None:
+        #     return Response({'error': 'Recipe name is required'}, status=status.HTTP_400_BAD_REQUEST)
+        # recipe = Recipe.objects.filter(name=recipe_name)
+        # if len(recipe) == 0:
+        #     return Response({'error': 'Recipe not found'}, status=status.HTTP_404_NOT_FOUND)
+        # serializer = RecipeSerializer(recipe[0])
+        # return Response(serializer.data, status=status.HTTP_200_OK)
+        return_data ={}
+        try:
+            recipe = Recipe.objects.get(id=recipe_id)
+        except Recipe.DoesNotExist:
+            return Response({'error': 'Recipe not found'}, status=404)
+
+        recipe_serializer = RecipeSerializer(recipe)
+        return_data = recipe_serializer.data
+        instructions = Instruction.objects.filter(recipe=recipe_id).order_by('step_number')
+        instruction_serializer = InstructionSerializer(instructions,many=True)
+        comments = Comment.objects.filter(recipe=recipe_id)
+        comment_serializer = CommentSerializer(comments,many=True)
+        diets = Diet.objects.filter(recipes=recipe_id)
+        diet_serializer = DietSerializer(diets,many=True)
+        cuisines = Cuisine.objects.filter(recipes=recipe_id)
+        cuisine_serializer = CuisineSerializer(cuisines,many=True)
+        ingredients = Ingredient.objects.filter(recipes=recipe_id)
+        ingredient_serializer= IngredientSerializer(ingredients,many=True)
+        return_data["ingredients"] = ingredient_serializer.data
+        return_data["cuisines"] = cuisine_serializer.data
+        return_data["diets"] = diet_serializer.data
+        return_data["instructions"] =instruction_serializer.data
+        return_data["comments"] = comment_serializer.data
+
+        return Response(return_data, status=200)
 
 
-class ShoppingList(View):
+class RetrieveCommentFilesView(APIView):
+    def get(self, request, comment_id):
+        comment_files = CommentFile.objects.filter(comment=comment_id)
+        files = []
+        for comment_file in comment_files:
+            file_path = comment_file.file.path
+            file = open(file_path, 'rb')
+            files.append(file)
+
+        response = FileResponse(files)
+        return response
+
+
+class RetrieveRecipeFilesView(APIView):
+    def get(self, request, recipe_id):
+        recipe_files = RecipeFile.objects.filter(recipe=recipe_id)
+        files = []
+        for recipe_file in recipe_files:
+            file_path = recipe_file.file.path
+            file = open(file_path, 'rb')
+            files.append(file)
+
+        response = FileResponse(files)
+        return response
+
+
+class RetrieveInstructionFilesView(APIView):
+    def get(self, request, instruction_id):
+        instruction_files = InstructionFile.objects.filter(recipe=instruction_id)
+        files = []
+        for instruction_file in instruction_files:
+            file_path = instruction_file.file.path
+            file = open(file_path, 'rb')
+            files.append(file)
+
+        response = FileResponse(files)
+        return response
+
+
+class AddToCartView(APIView):
+    def post(self, request, recipe_id):
+        user = request.user
+        try:
+            recipe = Recipe.objects.get(id=recipe_id)
+        except Recipe.DoesNotExist:
+            return Response(
+                {'Error': f'Recipe with id {recipe_id} does not exist'},
+                status=400)
+        shopping_list = ShoppingList.objects.get(user=user)
+        shopping_list.recipes.add(recipe)
+        shopping_list.save()
+        return Response({"message": "Successfully added recipe to the shopping list."}, status=201)
+
+class ReturnAllRecipes(ListAPIView):
+    serializer_class = RecipeSerializer
+    def get_queryset(self):
+        return Recipe.objects.all()
+
+class ShoppingLists(APIView):
+
     """
     This view is used to get the shopping list of a recipe.
     GET /recipes/shopping-list/
@@ -492,31 +600,36 @@ class ShoppingList(View):
     Note:
         The user must be logged in to use this view.
     """
+    permission_classes = [IsAuthenticated, IsTokenValid]
 
     def get(self, request):
-        username = request.GET.get('username', None)
-        if username is None:
-            return HttpResponseBadRequest("Username is required")
-        if not request.user.is_authenticated:
-            return HttpResponseForbidden("You must be logged in to use this view")
-        user = MyUser.objects.filter(username=username)
+
+        user = request.user
+
         if user is None:
             return HttpResponseBadRequest("User not found")
-        shopping_list = ShoppingList.objects.filter(user=user[0])
+
+        try:
+            shopping_list = ShoppingList.objects.get(user=user)
+        except ShoppingList.DoesNotExist:
+            return HttpResponseBadRequest("Shopping list not found", status=404)
+
         if shopping_list is None:
             return HttpResponseBadRequest("Shopping list not found")
-        recipes = shopping_list[0].recipes.all()
+        recipes = shopping_list.recipes.all()
         if recipes is None:
             return HttpResponseBadRequest("List of recipes are required")
-        result_json = {}
+        result_json = {'id': []}
+
         for recipe in recipes:
-            result_json[recipe.name] = {}
-            for ingredient in recipe.ingredients.all():
-                result_json[recipe.name][ingredient.name] = {
-                    'amount': ingredient.quantity,
-                    'unit_of_measure': ingredient.unit_of_measure
-                }
-            result_json[recipe.name]['servings'] = recipe.servings
+            result_json['id'].append(recipe.id)
+            # result_json[recipe.name] = {}
+            # for ingredient in recipe.ingredients.all():
+            #     result_json[recipe.name][ingredient.name] = {
+            #         'amount': ingredient.quantity,
+            #         'unit_of_measure': ingredient.unit_of_measure
+            #     }
+            # result_json[recipe.name]['servings'] = recipe.servings
         return JsonResponse(result_json, status=200)
 
 
@@ -533,7 +646,7 @@ class UpdateRecipeServings(APIView):
         The user must be logged in to use this view.
     """
     serializer_class = RecipeSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsTokenValid]
 
     def get_queryset(self):
         recipe_name = self.request.GET.get('recipe_name', None)
